Add recursive circuit public inputs:
It is done so that the recursion logic is agnostic to which data
structures are public inputs, and is the `InnerCircuit` & `OpsExecutor`
who internally define which targets are the public inputs and which are
the values that are used as public inputs for those targets.

For example at the current version, the `SchnorrPODGadget` which
implements the `InnerCircuitTrait`, returns an empty `Vec<F>=vec![]`;
and the `OpExecutorGadget` which implements the `OpsExecutorTrat`
returns a `Vec<F>` that comes from the values `Vec<Statement>`; so that
in the `RecursiveCircuit` the only public inputs appart from the
`VerifierData` are the ones of the `OpExecutorGadget`.
So, the public inputs in the PlonkyPOD are the values of the output list
of statements (Vec<Statement>).